Module zune_core::bit_depth

Image bit depth, information and manipulations

Enums

BitDepth

The image bit depth.

BitType

The underlying bit representation of the image

ByteEndian

Byte endianness of returned samples this is useful when the decoder returns samples which span more than one byte yet the type returned is &[u8]