Module zune_core::bytestream

A simple implementation of a bytestream reader and writer.

This module contains two main structs that help in byte reading and byte writing

Useful for a lot of image readers and writers, it’s put here to minimize code reuse

Structs

ZByteReader

An encapsulation of a byte stream reader

ZByteWriter

Encapsulates a simple Byte writer with support for Endian aware writes

Traits

ZReaderTrait

The underlying reader trait