zune_core::options

Struct DecoderOptions

pub struct DecoderOptions {
    max_width: usize,
    max_height: usize,
    out_colorspace: ColorSpace,
    max_scans: usize,
    deflate_limit: usize,
    flags: DecoderFlags,
    endianness: ByteEndian,
    hevc_max_mdat_size: usize,
    num_threads: u8,
}

Decoder options

Not all options are respected by decoders all decoders

Fields
max_width: usize

Maximum width for which decoders will not try to decode images larger than the specified width.

Default value: 16384

Respected by: all decoders

max_height: usize

Maximum height for which decoders will not try to decode images larger than the specified height

Default value: 16384

Respected by: all decoders

out_colorspace: ColorSpace

Output colorspace

The jpeg decoder allows conversion to a separate colorspace than the input.

I.e you can convert a RGB jpeg image to grayscale without first decoding it to RGB to get

Default value: ColorSpace::RGB

Respected by: jpeg

max_scans: usize

Maximum number of scans allowed for progressive jpeg images

Progressive jpegs have scans

Default value:100

Respected by: jpeg

deflate_limit: usize

Maximum size for deflate. Respected by all decoders that use inflate/deflate

flags: DecoderFlags

Boolean flags that influence decoding

endianness: ByteEndian

The byte endian of the returned bytes will be stored in in case a single pixel spans more than a byte

hevc_max_mdat_size: usize

Maximum MDAT size.

We read this to memory so thats why it is a configurable parameter

num_threads: u8

Number of threads used for decoding

Implementations

impl DecoderOptions

Initializers

pub fn new_safe() -> DecoderOptions

Create the decoder with options setting most configurable options to be their safe counterparts

This is the same as default option as default initializes options to the safe variant.

Note, decoders running on this will be slower as it disables platform specific intrinsics

pub fn new_fast() -> DecoderOptions

Create the decoder with options setting the configurable options to the fast counterparts

This enables platform specific code paths and enable use of unsafe

pub fn new_cmd() -> DecoderOptions

Create the decoder options with the following characteristics

Use unsafe paths.

Ignore error checksuming, e.g in png we do not confirm adler and crc in this mode

Enable fast intrinsics paths

impl DecoderOptions

Global options respected by all decoders

pub const fn max_width(&self) -> usize

Get maximum width configured for which the decoder should not try to decode images greater than this width

pub const fn max_height(&self) -> usize

Get maximum height configured for which the decoder should not try to decode images greater than this height

pub fn strict_mode(&self) -> bool

Return true whether the decoder should be in strict mode And reject most errors

pub const fn use_unsafe(&self) -> bool

Return true if the decoder should use unsafe routines where possible

pub fn set_max_width(self, width: usize) -> Self

Set maximum width for which the decoder should not try decoding images greater than that width

Arguments

width: The maximum width allowed

returns: DecoderOptions

pub fn set_max_height(self, height: usize) -> Self

Set maximum height for which the decoder should not try decoding images greater than that height

Arguments

height: The maximum height allowed

returns: DecoderOptions

pub fn set_use_unsafe(self, yes: bool) -> Self

Whether the routines can use unsafe platform specific intrinsics when necessary

Platform intrinsics are implemented for operations which the compiler can’t auto-vectorize, or we can do a marginably better job at it

All decoders with unsafe routines respect it.

Treat this with caution, disabling it will cause slowdowns but it’s provided for mainly for debugging use.

Respected by: png and jpeg(decoders with unsafe routines)

fn set_decoder_flags(self, flags: DecoderFlags) -> Self

pub fn set_strict_mode(self, yes: bool) -> Self

Set whether the decoder should be in standards conforming/ strict mode

This reduces the error tolerance level for the decoders and invalid samples will be rejected by the decoder

Arguments

yes:

returns: DecoderOptions

pub fn set_byte_endian(self, endian: ByteEndian) -> Self

Set the byte endian for which raw samples will be stored in in case a single pixel sample spans more than a byte.

The default is usually native endian hence big endian values will be converted to little endian on little endian systems,

and little endian values will be converted to big endian on big endian systems

Arguments

endian: The endianness to which to set the bytes to

returns: DecoderOptions

pub const fn byte_endian(&self) -> ByteEndian

Get the byte endian for which samples that span more than one byte will be treated

pub fn set_num_threads(self, num_threads: u8) -> Self

Set the number of threads used to decode images

This can be used e.g to implement threads used in heic tile decoding

pub const fn num_threads(&self) -> u8

Get the number of threads used to decode images

This can be used e.g to tell you how many threads the heic decoder will used when decoding tiles

impl DecoderOptions

PNG specific options

pub const fn inflate_get_confirm_adler(&self) -> bool

Whether the inflate decoder should confirm adler checksums

pub fn inflate_set_confirm_adler(self, yes: bool) -> Self

Set whether the inflate decoder should confirm adler checksums

pub const fn inflate_get_limit(&self) -> usize

Get default inflate limit for which the decoder will not try to decompress further

pub fn inflate_set_limit(self, limit: usize) -> Self

Set the default inflate limit for which decompressors relying on inflate won’t surpass this limit

pub const fn png_get_confirm_crc(&self) -> bool

Whether the inflate decoder should confirm crc 32 checksums

pub fn png_set_confirm_crc(self, yes: bool) -> Self

Set whether the png decoder should confirm CRC 32 checksums

pub fn png_set_add_alpha_channel(self, yes: bool) -> Self

Set whether the png decoder should add an alpha channel to images where possible.

For Luma images, it converts it to Luma+Alpha

For RGB images it converts it to RGB+Alpha

pub const fn png_get_add_alpha_channel(&self) -> bool

Return true whether the png decoder should add an alpha channel to images where possible

pub fn png_set_strip_to_8bit(self, yes: bool) -> Self

Whether the png decoder should reduce 16 bit images to 8 bit images implicitly.

Equivalent to png::Transformations::STRIP_16

pub const fn png_get_strip_to_8bit(&self) -> bool

Return a boolean indicating whether the png decoder should reduce 16 bit images to 8 bit images implicitly

pub const fn png_decode_animated(&self) -> bool

Return whether zune-image should decode animated images or whether we should just decode the first frame only

pub const fn png_set_decode_animated(self, yes: bool) -> Self

Set whether zune-image should decode animated images or whether we should just decode the first frame only

impl DecoderOptions

JPEG specific options

pub const fn jpeg_get_max_scans(&self) -> usize

Get maximum scans for which the jpeg decoder should not go above for progressive images

pub fn jpeg_set_max_scans(self, max_scans: usize) -> Self

Set maximum scans for which the jpeg decoder should not exceed when reconstructing images.

pub const fn jpeg_get_out_colorspace(&self) -> ColorSpace

Get expected output colorspace set by the user for which the image is expected to be reconstructed into.

This may be different from the

pub fn jpeg_set_out_colorspace(self, colorspace: ColorSpace) -> Self

Set expected colorspace for which the jpeg output is expected to be in

This is mainly provided as is, we do not guarantee the decoder can convert to all colorspaces and the decoder can change it internally when it sees fit.

impl DecoderOptions

Intrinsics support

These routines are compiled depending on the platform they are used, if compiled for a platform it doesn’t support,(e.g avx2 on Arm), it will always return false

pub fn use_sse2(&self) -> bool

Use SSE 2 code paths where possible

This checks for existence of SSE2 first and returns false if it’s not present

pub fn use_sse3(&self) -> bool

Use SSE 3 paths where possible

This also checks for SSE3 support and returns false if it’s not present

pub fn use_sse41(&self) -> bool

Use SSE4 paths where possible

This also checks for sse 4.1 support and returns false if it is not present

pub fn use_avx(&self) -> bool

Use AVX paths where possible

This also checks for AVX support and returns false if it’s not present

pub fn use_avx2(&self) -> bool

Use avx2 paths where possible

This also checks for AVX2 support and returns false if it’s not present

pub fn use_neon(&self) -> bool

impl DecoderOptions

JPEG_XL specific options

pub const fn jxl_decode_animated(&self) -> bool

Return whether zune-image should decode animated images or whether we should just decode the first frame only

pub const fn jxl_set_decode_animated(self, yes: bool) -> Self

Set whether zune-image should decode animated images or whether we should just decode the first frame only

impl DecoderOptions

HVEC decoding options

pub const fn hvec_use_apple_videotoolbox(&self) -> bool

Whether the decoder should use apple hardware decoding (videotoolbox) to decode heif/heic images.

pub const fn hvec_set_use_videotoolbox(self, yes: bool) -> Self

Set whether to use hardware decoding in heif/heic on apple devices

NB: This only affects decoding in macos its not considered for other os

pub const fn hevc_max_mdat_size(&self) -> usize

Return the size in bytes the maximum allowed size of the MDAT section in HEIC images, the section is read to memory so a cap is important

Default is 16 MB

pub fn set_hevc_max_mdat_size(self, max_size: usize) -> Self

Set the maximum size in bytes for the MDAT section for HEIC images.

The section is read into memory so important to have it with an upper limit