zune_core::options
Struct EncoderOptions

pub struct EncoderOptions { /* private fields */ }

Options shared by some of the encoders in the zune- family of image crates

Implementations

impl EncoderOptions

pub fn new( width: usize, height: usize, colorspace: ColorSpace, depth: BitDepth, ) -> EncoderOptions

Create new encode options

Arguments

width: Image width

height: Image height

colorspace: Image colorspaces

depth: Image depth

returns: EncoderOptions

pub const fn get_width(&self) -> usize

Get the width for which the image will be encoded in

pub fn get_height(&self) -> usize

Get height for which the image will be encoded in

returns: usize

Panics

If height is zero

pub const fn get_depth(&self) -> BitDepth

Get the depth for which the image will be encoded in

pub const fn get_quality(&self) -> u8

Get the quality for which the image will be encoded with

Lossy

Higher quality means some images take longer to write and are big but they look good

Lower quality means small images and low quality.

Lossless

High quality indicates more time is spent in making the file smaller

Low quality indicates less time is spent in making the file bigger

pub const fn get_colorspace(&self) -> ColorSpace

Get the colorspace for which the image will be encoded in

pub const fn get_effort(&self) -> u8

pub fn set_width(self, width: usize) -> Self

Set width for the image to be encoded

pub fn set_height(self, height: usize) -> Self

Set height for the image to be encoded

pub fn set_depth(self, depth: BitDepth) -> Self

Set depth for the image to be encoded

pub fn set_quality(self, quality: u8) -> Self

Set quality of the image to be encoded

Quality is clamped from 0..100

Quality means different options depending on the encoder, see get_quality

pub fn set_colorspace(self, colorspace: ColorSpace) -> Self

Set colorspace for the image to be encoded

pub fn set_num_threads(self, threads: u8) -> Self

Set the number of threads allowed for multithreaded encoding where supported

Zero means use a single thread

pub fn set_effort(self, effort: u8) -> Self

pub const fn get_num_threads(&self) -> u8

Return number of threads configured for multithreading where possible

This is used for multi-threaded encoders, currently only jpeg-xl

pub fn set_strip_metadata(self, yes: bool) -> Self

Set whether the encoder should remove metadata from the image

When set to true, supported encoders will strip away metadata from the resulting image. If set to false, where supported, encoders will not remove metadata from images

pub const fn strip_metadata(&self) -> bool

Whether or not the encoder should remove metadata from the image

The default value is false, and encoders that respect this try to preserve as much data as possible from one image to another

impl EncoderOptions

JPEG options

pub const fn jpeg_encode_progressive(&self) -> bool

Whether the jpeg encoder should encode the image in progressive mode

Default is false.

This may be used to create slightly smaller images at the cost of more processing time

pub const fn jpeg_optimized_huffman_tables(&self) -> bool

Whether the jpeg encoder should optimize huffman tables to create smaller files at the cost of processing time

Default is false.

pub fn set_jpeg_encode_progressive(self, yes: bool) -> Self

Set whether the jpeg encoder should encode the imagei in progressive mode

Default is false