zune_core::result
Enum DecodingResult

#[non_exhaustive]
pub enum DecodingResult {
    U8(Vec<u8>),
    U16(Vec<u16>),
    F32(Vec<f32>),
}

A simple enum that can hold decode results of most images

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

U8(Vec<u8>)

U16(Vec<u16>)

F32(Vec<f32>)

Implementations

impl DecodingResult

pub fn u8(self) -> Option<Vec<u8>>

Return the contents if the enum stores Vec<u8> or otherwise return None.

Useful for de-sugaring the result of a decoding operation into raw bytes

Example

use zune_core::result::DecodingResult;
let data = DecodingResult::U8(vec![0;100]);
// we know this won't fail because we created it with u8
assert!(data.u8().is_some());

let data = DecodingResult::U16(vec![0;100]);
// it should now return nothing since the type is u18
assert!(data.u8().is_none());

pub fn u16(self) -> Option<Vec<u16>>

Return the contents if the enum stores Vec<u16> or otherwise return None.

Useful for de-sugaring the result of a decoding operation into raw bytes

Example

use zune_core::result::DecodingResult;
let data = DecodingResult::U8(vec![0;100]);
// we know this will fail because we created it with u16
assert!(data.u16().is_none());

let data = DecodingResult::U16(vec![0;100]);
// it should now return something since the type is u16
assert!(data.u16().is_some());