Crate zune_core

Core routines shared by all libraries

This crate provides a set of core routines shared by the decoders and encoders under zune umbrella

It currently contains

A bytestream reader and writer with endian aware reads and writes

Colorspace and bit depth information shared by images

Image decoder and encoder options

A simple enum type to hold image decoding results.

This library is #[no_std] with alloc feature needed for defining Vec which we need for storing decoded bytes.

Features

no_std: Enables #[no_std] compilation support.

serde: Enables serializing of some of the data structures present in the crate

Modules

bit_depth

Image bit depth, information and manipulations

bytestream

A simple implementation of a bytestream reader and writer.

colorspace

Image Colorspace information and manipulation utilities.

log

options

Decoder and Encoder Options

result

Decoding results for images