zune_core::bytestream

Struct ZCursor

pub struct ZCursor<T: AsRef<[u8]>> { /* private fields */ }

Wraps an in memory buffer providing it with a Seek method but works in no_std environments

std::io::Cursor is available in std environments, but we also need support for no_std environments so this serves as a drop in replacement

Implementations

impl<T: AsRef<[u8]>> ZCursor<T>

pub fn new(buffer: T) -> ZCursor<T>

impl<T: AsRef<[u8]>> ZCursor<T>

pub fn skip(&mut self, num: usize)

Move forward num bytes from the current position.

It doesnโt check that position overflowed, new position may point past the internal buffer, all subsequent reads will either return an error or zero depending on the method called

pub fn rewind(&mut self, num: usize)

Move back num bytes from the current position

This saturates at zero, it can never be negative or wraparound when the value becomes too small